+++ flint/imager/wsclean.py
@@ -26,7 +26,6 @@ from typing import Any, Collection, Dict, List, NamedTuple, Optional, Tuple, Uni
 import numpy as np
 from fitscube.combine_fits import combine_fits
 
-from flint.calibrate.aocalibrate import AddModelOptions, add_model
 from flint.exceptions import CleanDivergenceError
 from flint.logging import logger
 from flint.ms import MS
@@ -61,6 +60,8 @@ class ImageSet(NamedTuple):
     """Model images.  """
     residual: Optional[List[Path]] = None
     """Residual images."""
+    source_list: Optional[Path] = None
+    """Path to a source list that accompanies the image data"""
 
 
 class WSCleanOptions(BaseOptions):
@@ -277,6 +278,7 @@ def _wsclean_output_callback(line: str) -> None:
         raise CleanDivergenceError(f"Clean divergence detected: {line}")
 
 
+# TODO: Update this function to also add int the source list
 def get_wsclean_output_names(
     prefix: str,
     subbands: int,
@@ -732,7 +734,6 @@ def run_wsclean_imager(
     move_hold_directories: Optional[Tuple[Path, Optional[Path]]] = None,
     make_cube_from_subbands: bool = True,
     image_prefix_str: Optional[str] = None,
-    calibrate_container: Optional[Path] = None,
 ) -> WSCleanCommand:
     """Run a provided wsclean command. Optionally will clean up files,
     including the dirty beams, psfs and other assorted things.
@@ -749,7 +750,6 @@ def run_wsclean_imager(
         move_hold_directories (Optional[Tuple[Path,Optional[Path]]], optional): The `move_directory` and `hold_directory` passed to the temporary context manager. If None no `hold_then_move_into` manager is used. Defaults to None.
         make_cube_from_subbands (bool, optional): Form a single FITS cube from the set of sub-band images wsclean produces. Defaults to False.
         image_prefix_str (Optional[str], optional): The name used to search for wsclean outputs. If None, it is guessed from the name and location of the MS. Defaults to None.
-        calibrate_container (Optional[Path], optional): Path to the ``aocalibrate`` container with the ``addmodel`` program. Should this be provided and the ``wsclean -save-source-list`` be used, then the model will be predicted at full channel resolution
 
     Returns:
         WSCleanCommand: The executed wsclean command with a populated imageset properter.
@@ -803,24 +803,6 @@ def run_wsclean_imager(
         #         prefix=prefix, output_type=output_type, ignore_mfs=False
         #     )
 
-    if calibrate_container and wsclean_cmd.options.save_source_list:
-        logger.info("Predicting the wsclean clean components SEDs")
-        source_list_path = get_wsclean_output_source_list_path(
-            name_path=prefix, pol=None
-        )
-        assert source_list_path.exists(), f"{source_list_path=} does not exist"
-        add_model_options = AddModelOptions(
-            model_path=source_list_path,
-            ms_path=wsclean_cmd.ms.path,
-            mode="c",
-            datacolumn="MODEL_DATA",
-        )
-        add_model(
-            add_model_options=add_model_options,
-            container=calibrate_container,
-            remove_datacolumn="MODEL_DATA",
-        )
-
     imageset = get_wsclean_output_names(
         prefix=prefix,
         subbands=wsclean_cmd.options.channels_out,
@@ -828,6 +810,14 @@ def run_wsclean_imager(
         output_types=("image", "residual"),
         check_exists_when_adding=True,
     )
+
+    if wsclean_cmd.options.save_source_list:
+        logger.info("Attaching the wsclean clean components SEDs")
+        source_list_path = get_wsclean_output_source_list_path(
+            name_path=prefix, pol=None
+        )
+        assert source_list_path.exists(), f"{source_list_path=} does not exist"
+        imageset = imageset.with_options(source_list=source_list_path)
 
     if make_cube_from_subbands:
         imageset = combine_subbands_to_cube(
@@ -845,7 +835,6 @@ def wsclean_imager(
     ms: Union[Path, MS],
     wsclean_container: Path,
     update_wsclean_options: Optional[Dict[str, Any]] = None,
-    calibrate_container: Optional[Path] = None,
 ) -> WSCleanCommand:
     """Create and run a wsclean imager command against a measurement set.
 
@@ -853,7 +842,6 @@ def wsclean_imager(
         ms (Union[Path,MS]): Path to the measurement set that will be imaged
         wsclean_container (Path): Path to the container with wsclean installed
         update_wsclean_options (Optional[Dict[str, Any]], optional): Additional options to update the generated WscleanOptions with. Keys should be attributes of WscleanOptions. Defaults to None.
-        calibrate_container (Optional[Path], optional): Patht to the aocalibrate container with ``addmodel``. If not None and ``wsclean -save-source-list` is used the model will be predicted at full channel resolution
 
     Returns:
         WSCleanCommand: _description_
@@ -873,7 +861,6 @@ def wsclean_imager(
         ms=ms,
         wsclean_options=wsclean_options,
         container=wsclean_container,
-        calibrate_container=calibrate_container,
     )
 
     return wsclean_cmd